Added VESA 800x600x32 terminal support and refactored standard library

+++ include/builtins/string.h
@@ -5,7 +5,7 @@
 
 #define USE_BUILTIN_STRING
 
-static inline void* memset_slow(void* addr, int32 val, int32 len) {
+static inline void* memset_slow(void* addr, size_t val, size_t len) {
 	byte* p = (byte*)addr;
 
 	while (len--)
@@ -14,7 +14,7 @@ static inline void* memset_slow(void* addr, int32 val, int32 len) {
 	return addr;
 }
 
-static inline void* memcpy_slow(void* dest, const void* src, int32 len) {
+static inline void* memcpy_slow(void* dest, const void* src, size_t len) {
 	byte* p;
 
 	byte* q;
@@ -29,7 +29,7 @@ static inline void* memcpy_slow(void* dest, const void* src, int32 len) {
 	return dest;
 }
 
-static inline int32 strcmp_slow(const byte* s1, const byte* s2) {
+static inline size_t strcmp_slow(const byte* s1, const byte* s2) {
 	int32 r, c1, c2;
 
 	do {
@@ -43,7 +43,7 @@ static inline int32 strcmp_slow(const byte* s1, const byte* s2) {
 	return r;
 }
 
-static inline int32 memcmp_slow(const void* p1, const void* p2, int32 len) {
+static inline size_t memcmp_slow(const void* p1, const void* p2, size_t len) {
 	int32 r, i;
 
 	byte* q1;
@@ -60,7 +60,7 @@ static inline int32 memcmp_slow(const void* p1, const void* p2, int32 len) {
 	return r;
 }
 
-static inline int32 strlen_slow(const byte* p) {
+static inline size_t strlen_slow(const byte* p) {
 	int32 len = 0;
 
 	while (*p++)
@@ -69,7 +69,7 @@ static inline int32 strlen_slow(const byte* p) {
 	return len;
 }
 
-static inline int32 strncmp_slow(const byte* s1, const byte* s2, int32 len) {
+static inline size_t strncmp_slow(const byte* s1, const byte* s2, size_t len) {
 	int32 r, c1, c2;
 
 	if (len <= 0)
@@ -104,19 +104,19 @@ static inline int32 strncmp_slow(const byte* s1, const byte* s2, int32 len) {
 
 #ifdef USE_BUILTIN_STRING
 
-static inline void* memset_builtin(void* addr, int val, uint64 len) {
+static inline void* memset_builtin(void* addr, int val, size_t len) {
 	return __builtin_memset(addr, val, len);
 }
 
-static inline void* memcpy_builtin(void* dest, const void* src, uint64 len) {
+static inline void* memcpy_builtin(void* dest, const void* src, size_t len) {
 	return __builtin_memcpy(dest, src, len);
 }
 
-static inline uint64 strcmp_builtin(const byte* s1, const byte* s2) {
+static inline size_t strcmp_builtin(const byte* s1, const byte* s2) {
 	return __builtin_strcmp(s1, s2);
 }
 
-static inline uint64 memcmp_builtin(const void* p1, const void* p2, uint64 len) {
+static inline size_t memcmp_builtin(const void* p1, const void* p2, size_t len) {
 	if (__builtin_constant_p(len) && len == 2) {
 		uint16* q1 = p1, *q2 = p2;
 
@@ -132,11 +132,11 @@ static inline uint64 memcmp_builtin(const void* p1, const void* p2, uint64 len) 
 	return __builtin_memcmp(p1, p2, len);
 }
 
-static inline uint64 strlen_builtin(const byte* p) {
+static inline size_t strlen_builtin(const byte* p) {
 	return __builtin_strlen(p);
 }
 
-static inline uint64 strncmp_builtin(const byte* s1, const byte* s2, uint64 len) {
+static inline size_t strncmp_builtin(const byte* s1, const byte* s2, size_t len) {
 	return __builtin_strncmp(s1, s2, len);
 }
